更新项目：添加制裁地址检查和合规管理功能
1. 在所有涉及地址的操作中添加制裁地址检查，包括：
   - 资产创建
   - 存款
   - 赎回
   - 转账
   - 用户注册

2. 实现合规管理页面，允许监管者：
   - 手动添加制裁地址
   - 同步制裁名单
   - 查看制裁名单

3. 改进系统功能：
   - 修复资产管理界面网络错误
   - 修复登录失败问题
   - 修复余额显示为0的问题
   - 转换价格单位从美分到美元
   - 添加日志记录功能，记录与制裁地址相关的操作

4. 优化用户体验：
   - 改进UI布局，增加间距
   - 提供明确的错误信息
   - 确保前后端同步修改

+++ backend/internal/database/asset.go
@@ -10,7 +10,7 @@ type Asset struct {
 	AssetID      string
 	Name         string
 	Symbol       string
-	TotalValue   uint64
+	TotalValue   float64
 	TotalTokens  uint64
 	TokenAddress string
 	IsActive     bool
@@ -58,7 +58,7 @@ func GetAssetByID(assetID string) (Asset, error) {
 }
 
 // UpdateAssetValue 更新资产价值
-func UpdateAssetValue(assetID string, value uint64, isDeposit bool) error {
+func UpdateAssetValue(assetID string, value float64, isDeposit bool) error {
 	var err error
 	if isDeposit {
 		_, err = DB.Exec(
